Analysis conditioning on acquisition variables
As requested, I conditioned on other image acquisition variables provided, namely the charge used in mAs, and an estimate of patient diameter. A few different visualizations were tried in the conditionByCat script. Also, for expediency the prediction scores were saved as an additional variable in the info csv. Lastly, I started using the seaborn package in this evaluation, and it has some interesting abilities I will continue to use in the future.